Reduce duplicated code from DAM
