+++ packages/api/cms-api/src/dam/images/images.constants.ts
@@ -1,0 +1,10 @@
+const WEBP = "image/webp";
+const PNG = "image/png";
+const JPEG = "image/jpeg";
+const GIF = "image/gif";
+
+export const BASIC_TYPES = [JPEG, PNG, GIF];
+
+export const MODERN_TYPES = [/* AVIF, */ WEBP];
+
+export const ALL_TYPES = [...BASIC_TYPES, ...MODERN_TYPES];
